refactor: :fire: remove test files
